Moved the form component to coco subfolder
